+++ breeze-dsl/breeze-dsl-mermaid/src/main/kotlin/com/windea/breezeframework/dsl/mermaid/MermaidPieChartDsl.kt
@@ -71,9 +71,9 @@ class MermaidPieChartSection @PublishedApi internal constructor(
 ) : MermaidPieChartDslElement, WithUniqueId {
 	override val id: String get() = key
 	
-	override fun equals(other: Any?) = equalsBySelectId(this, other) { id }
+	override fun equals(other: Any?) = equalsByOne(this, other) { id }
 	
-	override fun hashCode() = hashCodeBySelectId(this) { id }
+	override fun hashCode() = hashCodeByOne(this) { id }
 	
 	override fun toString(): String {
 		return "${key.wrapQuote(quote)}: $value"