+++ breeze-dsl/breeze-dsl-mermaid/src/main/kotlin/com/windea/breezeframework/dsl/mermaid/MermaidSequenceDiagramDsl.kt
@@ -73,9 +73,9 @@ class MermaidSequenceDiagramParticipant @PublishedApi internal constructor(
 	
 	override val id: String get() = alias ?: name
 	
-	override fun equals(other: Any?) = equalsBySelectId(this, other) { id }
+	override fun equals(other: Any?) = equalsByOne(this, other) { id }
 	
-	override fun hashCode() = hashCodeBySelectId(this) { id }
+	override fun hashCode() = hashCodeByOne(this) { id }
 	
 	override fun toString(): String {
 		val aliasSnippet = alias?.let { "$it as " }.orEmpty()