+++ breeze-dsl/breeze-dsl-puml/src/main/kotlin/com/windea/breezeframework/dsl/puml/PumlSequenceDiagramDsl.kt
@@ -49,9 +49,9 @@ class PumlSequenceDiagramParticipant @PublishedApi internal constructor(
 	
 	override val id: String get() = alias ?: name
 	
-	override fun equals(other: Any?) = equalsBySelectId(this, other) { id }
+	override fun equals(other: Any?) = equalsByOne(this, other) { id }
 	
-	override fun hashCode() = hashCodeBySelectId(this) { id }
+	override fun hashCode() = hashCodeByOne(this) { id }
 	
 	override fun toString(): String {
 		val orderSnippet = order?.let { " order $order" }.orEmpty()