+++ breeze-dsl/breeze-dsl-sequence/src/main/kotlin/com/windea/breezeframework/dsl/sequence/SequenceDiagramDsl.kt
@@ -80,9 +80,9 @@ class SequenceDiagramParticipant @PublishedApi internal constructor(
 	
 	override val id: String get() = alias ?: name
 	
-	override fun equals(other: Any?) = equalsBySelectId(this, other) { id }
+	override fun equals(other: Any?) = equalsByOne(this, other) { id }
 	
-	override fun hashCode() = hashCodeBySelectId(this) { id }
+	override fun hashCode() = hashCodeByOne(this) { id }
 	
 	override fun toString(): String {
 		val aliasSnippet = alias?.let { "as $it" }.orEmpty()