+++ CHANGELOG.md
@@ -219,6 +219,7 @@
 
 * [X] 更新版本。
 * [X] 清理注解。
+* [X] breeze-core 更新`DataClassExtensions`，将相关扩展提取为`AnyExtensions`，并提供`smartEquals`等方法。
 
 # 待办